+++ lib/account module/home_screen.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) => Scaffold(
+        body: CustomScrollView(
+          slivers: [
+            SliverAppBar(
+              title: Text('hello'),
+              expandedHeight: 400,
+              flexibleSpace: FlexibleSpaceBar(
+                background: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Text('Cash'),
+                    Text(r'$400,000'),
+                    Text(r'Today: -$30,000')
+                  ],
+                ),
+              ),
+            ),
+            SliverToBoxAdapter(
+              child: SizedBox(
+                height: 2000,
+              ),
+            )
+          ],
+        ),
+      );
+}