+++ lib/components/loans_view.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LoansView extends StatelessWidget {
-  const LoansView({super.key});
-
-  @override
-  Widget build(BuildContext context) => const Text('Loans View');
-}