+++ lib/components/sign_in_view.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SignInView extends StatelessWidget {
-  const SignInView({super.key});
-
-  @override
-  Widget build(BuildContext context) => const Text('Sign In');
-}