+++ lib/shared/route_names.dart
@@ -6,33 +6,14 @@ class Routes {
   static const onboarding = '/onboarding';
   static const splash = '/splash';
 
+  //NavBar Routes
+  static const dashboard = '/dashboard';
+  static const records = '/records';
+  static const other = '/other';
+
   //Settings Routes
   static const CHOOSE_CURRENCY = '/chooseCurrency';
   static const CHOOSE_DEFAULT_CURRENCY = '/chooseDefaultCurrency';
   static const EDIT_CURRENCY = '/editCurrency';
   static const CREATE_NEW_CURRENCY = '/createNewCurrency';
-
-  //Navrail Route
-  static const DASHBOARD_ROUTE = '/dashboard';
 }
-
-//Auth Routes
-const SIGN_IN_ROUTE = '/signIn';
-const SIGN_UP_ROUTE = '/signUp';
-const CONFIRM_EMAIL_ROUTE = '/confirmEmail';
-const FORGOT_PASSWORD_ROUTE = '/forgotPassword';
-
-//Bottom NavBar routes
-const HOME_ROUTE = '/';
-const ACCOUNT_ROUTE = '/account';
-const TRANSACTIONS_ROUTE = '/records';
-const REPORT_ROUTE = '/report';
-const MORE_ROUTE = '/more';
-
-//Navrail Routes
-const DASHBOARD_ROUTE = '/dashboard';
-const TIMELINE_ROUTE = '/timeline';
-const SETTING_ROUTE = '/setting';
-
-//Settings Routes
-const DEFAULT_CURRENCY_ROUTE = 'defaultCurrency';