+++ lib/util/paths.dart
@@ -1,8 +1,0 @@
-// ignore_for_file: constant_identifier_names
-
-class Paths {
-  //static final sqlite3_path = '';
-  static const APP_LOGO_PATH_LIGHT = 'assets/images/app_logo.png';
-  static const APP_LOGO_PATH_DARK = 'assets/images/app_logo.png';
-  static const PRELOADED_CURRENCY_JSON = 'assets/json/loaded_currencies.json';
-}